W5D4/requirements.txt
streamlit
requests
python-dotenv

plotly

fastapi[standard]
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
fastapi-users[sqlalchemy]
alembic
unstructured
python-docx
langchain
langchain-community
langchain-core
langchain-groq
opencv-python-headless
qdrant-client
sentence-transformers
python-dotenv
chromadb
pypdf
beautifulsoup4
tiktoken
faiss-cpu
numpy
pandas
pydantic-settings
python-magic
pytesseract
pdf2image


asyncpg
aiohttp
fpdf
markdown
